Move flower info to new component
