+++ frontend/src/components/FlowerModal.jsx
@@ -1,34 +1,11 @@
 import { Modal, Button, Tabs, Tab } from 'react-bootstrap'
 import { useTranslation } from 'react-i18next'
-import VisibilityButton from './VisibilityButton'
-import ModifyFlowerForm from './ModifyFlowerForm'
+import FlowerInfoTab from './FlowerInfoTab'
 import FlowerImageTab from './image/FlowerImageTab'
-import { useState } from "react"
+
 
 const FlowerModal = ({ show, handleClose, flower, deleteFlower, updateFlower, modifyFlower }) => {
   const { t } = useTranslation()
-  const [isModifyFormVisible, setIsModifyFormVisible] = useState(false)
-
-  const handleFlowerDelete = (flower) => {
-    if (deleteFlower) {
-      deleteFlower(flower)
-    }
-    handleClose()
-  }
-
-  const handleFormVisibility = () => {
-    setIsModifyFormVisible((prev) => !prev)
-  }
-
-  const addedTime = (flower) => {
-    let addedTime = new Date(flower.added_time)
-
-    let date = addedTime.toLocaleDateString('fi')
-    let time = addedTime.toLocaleTimeString('fi', { hour: '2-digit', minute: '2-digit' })
-    let addedTimeStr = `${date} ${time}`
-
-    return addedTimeStr
-  }
 
   const isGrower = Boolean(deleteFlower && updateFlower && modifyFlower)
 
@@ -38,50 +15,10 @@ const FlowerModal = ({ show, handleClose, flower, deleteFlower, updateFlower, mo
         <Modal.Title>{flower.name}</Modal.Title>
       </Modal.Header>
       <Modal.Body>
-        <Tabs
-          defaultActiveKey="info"
-          id="uncontrolled-tab-example"
-          className="mb-3"
-          mountOnEnter={true}
-          unmountOnExit={true}
-          >
+        <Tabs defaultActiveKey="info" className="mb-3" mountOnEnter={true} unmountOnExit={true}>
           <Tab eventKey="info" title={t('menu.info')}>
             <div>
-              {isGrower && isModifyFormVisible ? (
-                  <div>
-                    <ModifyFlowerForm 
-                      flower={flower} 
-                      modifyFlower={modifyFlower} 
-                      handleFlowerModify={updateFlower}
-                      handleFormVisibility={handleFormVisibility}
-                    />
-                  </div> 
-                ) : (
-                  <div>
-                    <h3>{t('menu.info')}</h3>
-                    <p>{t('flower.data.name')}: {flower.name}</p>
-                    <p>{t('flower.data.latinname')}: {flower.latin_name}</p>
-                    <p>{t('flower.data.addedtime')}: {addedTime(flower)}</p>
-                    <p>{t('flower.data.site')}: {flower.site_name}</p>
-                    <p>{t('flower.data.qty')}: {flower.quantity}</p>
-                  </div>
-                )}
-              {isGrower ?
-              <p>{t('flower.visible.long')}: {flower.visible 
-                    ? t('flower.visible.true') 
-                    : t('flower.visible.false')}
-              <VisibilityButton flower={flower} updateFlower={updateFlower}/>
-              </p> : <></>}
-              {deleteFlower && (
-                <button id="deleteFlowerButton" onClick={() => handleFlowerDelete(flower)}>
-                  {t('button.delete')}
-                </button>
-              )}
-              {isGrower && !isModifyFormVisible && (
-                <button id="modifyFlowerButton" onClick={handleFormVisibility}>
-                  {t('button.modify')}
-                </button>
-              )}
+              <FlowerInfoTab isGrower={isGrower} flower={flower} deleteFlower={deleteFlower} updateFlower={updateFlower} modifyFlower={modifyFlower} handleClose={handleClose}/>
             </div>
           </Tab>
           <Tab eventKey="images" title={t('menu.images')}>
